move cast signature to expr crate
